remove refresh logic
